move EW JS API object files directly under __mocks__ folder
