+++ test/examples/forward-devicetype/main.test.js
@@ -1,5 +1,5 @@
 import onClientRequest from "../../../src/examples/forward-devicetype/main";
-import Request from "../../../__mocks__/object/request";
+import Request from "../../../__mocks__/request";
 
 describe('onClientRequest should modify forward path based on device type to point to device specific content', () => {
