+++ test/examples/response-manipulation-stream/main.test.js
@@ -1,7 +1,7 @@
 import {httpRequest, HttpResponse, mock_HttpResponse_json} from "../../../__mocks__/http-request";
 import {createResponse} from "../../../__mocks__/create-response";
 import {responseProvider} from "../../../src/examples/response-manipulation-stream/main";
-import Request from "../../../__mocks__/object/request";
+import Request from "../../../__mocks__/request";
 import {TextDecoderStream, TextEncoderStream} from "../../../__mocks__/text-encode-transform";
