+++ test/examples/traffic-allow-list/main.test.js
@@ -1,5 +1,5 @@
 import onClientRequest from "../../../src/examples/traffic-allow-list/main";
-import Request from "../../../__mocks__/object/request";
+import Request from "../../../__mocks__/request";
 
 describe('onClientRequest should modify allow or deny message depending on country of end user', () => {
